Cleanup - removed ./out dir - moved middleware and handlers into server dir - renamed data package to model
