--- api/data/movies/models.go
@@ -1,7 +1,0 @@
-package data
-
-type Movie struct {
-	ID      string `json:"id"`
-	Title   string `json:"title"`
-	Watched bool   `json:"watched"`
-}
