--- api/data/movies/movies.json
@@ -1,22 +1,0 @@
-[
-{
-"id": "0",
-"title": "Shawshank Redemption",
-"watched": false
-},
-{
-"id": "1",
-"title": "The Godfather",
-"watched": false
-},
-{
-"id": "2",
-"title": "The Godfather: Part II",
-"watched": false
-},
-{
-"id": "3",
-"title": "The Dark Knight",
-"watched": false
-}
-]
