--- api/handlers/healthz.go
@@ -1,11 +1,0 @@
-package handlers
-
-import (
-	"github.com/gin-gonic/gin"
-)
-
-func HealthzHandler(c *gin.Context) {
-	c.JSON(200, gin.H{
-		"status": "healthy",
-	})
-}
