+++ api/main.go
@@ -6,9 +6,13 @@ import (
 	server "github.com/kindaqt/movies/api/server"
 )
 
+const (
+	port string = ":8080"
+)
+
 func main() {
 	fmt.Println("Starting Movies API")
 
 	router := server.Router()
-	router.Run(":8081")
+	router.Run(port)
 }
