--- api/middleware/cors.go
@@ -1,26 +1,0 @@
-package middleware
-
-import (
-	"fmt"
-
-	"github.com/gin-gonic/gin"
-)
-
-var CLIENT string = "http://localhost:3000"
-
-func CORSMiddleware() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		fmt.Println("HELLO")
-		c.Writer.Header().Set("Access-Control-Allow-Origin", CLIENT)
-		c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
-		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
-		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, DELETE")
-
-		if c.Request.Method == "OPTIONS" {
-			c.AbortWithStatus(204)
-			return
-		}
-
-		c.Next()
-	}
-}
