+++ api/model/movies/models.go
@@ -1,0 +1,7 @@
+package model
+
+type Movie struct {
+	ID      string `json:"id"`
+	Title   string `json:"title"`
+	Watched bool   `json:"watched"`
+}
