+++ api/server/handlers/healthz.go
@@ -1,0 +1,11 @@
+package handlers
+
+import (
+	"github.com/gin-gonic/gin"
+)
+
+func HealthzHandler(c *gin.Context) {
+	c.JSON(200, gin.H{
+		"status": "healthy",
+	})
+}
