+++ api/server/handlers/healthz_test.go
@@ -1,0 +1,28 @@
+package handlers
+
+import (
+	"encoding/json"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/gin-gonic/gin"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestHealthzHandler(t *testing.T) {
+
+	w := httptest.NewRecorder()
+	c, _ := gin.CreateTestContext(w)
+
+	HealthzHandler(c)
+
+	// Assert Status Code
+	assert.Equal(t, http.StatusOK, w.Code)
+
+	var expectedResponse map[string]string = map[string]string{"status": "healthy"}
+	var response map[string]string
+	err := json.Unmarshal([]byte(w.Body.String()), &response)
+	assert.Equal(t, err, nil)
+	assert.Equal(t, expectedResponse, response)
+}
